fix(sdk,core): break the types/routing package cycle — SDK owns the route type declarations
The import-cycle ratchet caught a NEW package cycle introduced by T19's
APIRoute unification: sdk/types/index re-exported APIRoute +
PluginContextLite through ../routing, whose barrel pulls @bakin/core/routing
(value imports incl. the dispatcher) — while core/plugin-types has always
imported @makinbakin/sdk/types. Mutual package dependency.

Break (verified variant of the planned move): the pure, CONTEXT-FREE route
declarations — generic APIRoute (C = unknown), body/response specs,
ParsedInput, HttpStatus — now live in a leaf module
packages/sdk/src/types/api-route.ts importing only sibling SDK type modules.
core/routing/types.ts re-exports them and keeps ONLY the tier-specific
contexts (RouteContext/PluginContextLite/CoreContext), exporting APIRoute as
an alias with the core-tier RouteContext default — one declaration, pinned
mutually assignable by the existing plugin-type-contract test.
PluginContextLite deliberately did NOT move into the SDK leaf: RouteContext's
members are core-tier (AgentRuntimeAdapter et al.), and an SDK-tier Lite
would break handler assignability against defineRoute's core-tier ctx (the
SDK runtime type's channels-required dishonesty is already flagged for PR 3).
It remains published via @makinbakin/sdk/routing, which nothing imported it
from @makinbakin/sdk/types anyway (grep-verified). BakinPlugin.routes' inline
../routing import repointed to the leaf with ctx as any — same documented
rationale as activate()'s ctx; authoring-time typing rides defineRoute.

Resulting import graph: sdk/types is a leaf; core → sdk/types;
sdk/routing → core/routing. check:cycles: 0 new, allowlist unchanged.
